requirements.txt
faiss_gpu==1.7.1
numpy~=1.21.2
Pillow==8.4.0
scikit_learn==1.0.1
torchvision~=0.11.2
tqdm==4.48.2

scikit-learn~=1.0.1

h5py~=2.10.0